GetCmdEstimateSwapExactAmountOut to osmocli/swaprouter
